primenumber revamp working
