+++ Features/oracle-primenumber/build.gradle
@@ -4,15 +4,16 @@ buildscript {
 
     ext {
         corda_release_group = constants.getProperty("cordaReleaseGroup")
+        corda_core_release_group = constants.getProperty("cordaCoreReleaseGroup")
         corda_release_version = constants.getProperty("cordaVersion")
-        corda_core_release_group =  constants.getProperty("cordaCoreReleaseGroup")
         corda_core_release_version = constants.getProperty("cordaCoreVersion")
         corda_gradle_plugins_version = constants.getProperty("gradlePluginsVersion")
+        kotlin_version = constants.getProperty("kotlinVersion")
         junit_version = constants.getProperty("junitVersion")
         quasar_version = constants.getProperty("quasarVersion")
         log4j_version = constants.getProperty("log4jVersion")
         slf4j_version = constants.getProperty("slf4jVersion")
-        corda_platform_version = constants.getProperty("platformVersion")
+        corda_platform_version = constants.getProperty("platformVersion").toInteger()
     }
 
     repositories {
@@ -33,55 +34,80 @@ allprojects {
     apply from: "${rootProject.projectDir}/repositories.gradle"
     apply plugin: 'java'
 
-    tasks.withType(JavaCompile) {
-        options.compilerArgs << "-parameters" // Required for shell commands.
+    repositories {
+        mavenLocal()
+        jcenter()
+        mavenCentral()
+        maven { url 'https://software.r3.com/artifactory/corda' }
+        maven { url 'https://jitpack.io' }
     }
 
+    tasks.withType(JavaCompile) {
+        options.compilerArgs << "-parameters" // Required by Corda's serialisation framework.
+    }
 
+    jar {
+        // This makes the JAR's SHA-256 hash repeatable.
+        preserveFileTimestamps = false
+        reproducibleFileOrder = true
+        duplicatesStrategy = DuplicatesStrategy.EXCLUDE
+    }
 }
 
 apply plugin: 'net.corda.plugins.quasar-utils'
 apply plugin: 'net.corda.plugins.cordapp'
 apply plugin: 'net.corda.plugins.cordformation'
 
+sourceSets {
+    main {
+        resources {
+            srcDir rootProject.file("config/dev")
+        }
+    }
+}
+
 cordapp {
     targetPlatformVersion corda_platform_version.toInteger()
     minimumPlatformVersion corda_platform_version.toInteger()
     info {
-        name "Oracle Example"
+        name "Oracle Primenumber"
         vendor "Corda Open Source"
     }
 }
 
 dependencies {
-    testCompile "junit:junit:$junit_version"
-
     // Corda dependencies.
+    cordaRuntime "$corda_release_group:corda:$corda_release_version"
     cordaCompile "$corda_core_release_group:corda-core:$corda_core_release_version"
+    cordaCompile "$corda_release_group:corda-node-api:$corda_release_version"
     cordaCompile "$corda_release_group:corda-jackson:$corda_release_version"
     cordaCompile "$corda_release_group:corda-rpc:$corda_release_version"
-    cordaCompile "$corda_release_group:corda-node-api:$corda_release_version"
-    cordaRuntime "$corda_release_group:corda:$corda_release_version"
-
-    testCompile "$corda_release_group:corda-node-driver:$corda_release_version"
-    cordaRuntime "org.apache.logging.log4j:log4j-slf4j-impl:$log4j_version"
 
     // CorDapp dependencies.
-    cordapp project(":base")
-    cordapp project(":client")
-    cordapp project(":service")
+    cordapp project(":workflows")
+    cordapp project(":contracts")
+
+    cordaRuntime "org.apache.logging.log4j:log4j-slf4j-impl:$log4j_version"
+    cordaCompile "org.apache.logging.log4j:log4j-slf4j-impl:${log4j_version}"
+    cordaCompile "org.apache.logging.log4j:log4j-web:${log4j_version}"
+    cordaCompile "org.slf4j:jul-to-slf4j:$slf4j_version"
+
+    testCompile "junit:junit:$junit_version"
+    testCompile "$corda_release_group:corda-node-driver:$corda_release_version"
 }
 
 task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
+
     nodeDefaults {
         projectCordapp {
             deploy = false
         }
-	runSchemaMigration = true
+        runSchemaMigration = true
     }
+
     node {
         name "O=Notary,L=London,C=GB"
-        notary = [validating : false]
+        notary = [validating: false]
         p2pPort 10002
         rpcSettings {
             address("localhost:10003")
@@ -89,6 +115,7 @@ task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
         }
         cordapps = []
     }
+
     node {
         name "O=PartyA,L=London,C=GB"
         p2pPort 10005
@@ -96,10 +123,11 @@ task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
             address("localhost:10006")
             adminAddress("localhost:10046")
         }
-        cordapp project(':base')
-        cordapp project(':client')
-        rpcUsers = [[ user: "user1", "password": "test", "permissions": ["ALL"]]]
+        cordapp project(":contracts")
+        cordapp project(":workflows")
+        rpcUsers = [[user: "user1", "password": "test", "permissions": ["ALL"]]]
     }
+
     node {
         name "O=Oracle,L=New York,C=US"
         p2pPort 10008
@@ -107,8 +135,8 @@ task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
             address("localhost:10009")
             adminAddress("localhost:10049")
         }
-        cordapp project(':base')
-        cordapp project(':service')
-        rpcUsers = [[ user: "user1", "password": "test", "permissions": ["ALL"]]]
+        cordapp project(":contracts")
+        cordapp project(":workflows")
+        rpcUsers = [[user: "user1", "password": "test", "permissions": ["ALL"]]]
     }
 }
