+++ Features/oracle-primenumber/gradle.properties
@@ -1,3 +1,3 @@
-name=Test
-group=net.corda.examples.oracle
+name=Oracle Primenumber
+group=com.oracle
 version=1.0
