--- Features/oracle-primenumber/service/build.gradle
@@ -1,29 +1,0 @@
-apply plugin: 'net.corda.plugins.cordapp'
-apply plugin: 'net.corda.plugins.quasar-utils'
-
-cordapp {
-    targetPlatformVersion corda_platform_version.toInteger()
-    minimumPlatformVersion corda_platform_version.toInteger()
-    workflow {
-        name "Oracle Example"
-        vendor "Corda Open Source"
-        licence "Apache License, Version 2.0"
-        versionId 1
-    }
-}
-
-dependencies {
-    testCompile "junit:junit:$junit_version"
-
-    // Corda dependencies.
-    cordaCompile "$corda_core_release_group:corda-core:$corda_core_release_version"
-    cordaCompile "$corda_release_group:corda-jackson:$corda_release_version"
-    cordaCompile "$corda_release_group:corda-rpc:$corda_release_version"
-    cordaCompile "$corda_release_group:corda-node-api:$corda_release_version"
-
-    testCompile "$corda_release_group:corda-node-driver:$corda_release_version"
-
-    // CorDapp dependencies.
-    cordapp project(":base")
-    cordapp project(":client")
-}
