+++ Features/oracle-primenumber/settings.gradle
@@ -1,3 +1,3 @@
-include 'base'
-include 'service'
-include 'client'
+include 'contracts'
+include 'workflows'
+
